+++ packages/cketh/src/orchestrator.canister.spec.ts
@@ -1,6 +1,6 @@
 import type { ActorSubclass } from "@dfinity/agent";
 import { Principal } from "@dfinity/principal";
-import { mock } from "jest-mock-extended";
+import { mock } from "vitest-mock-extended";
 import type {
   _SERVICE as CkETHOrchestratorService,
   ManagedCanisters,