+++ test-setup.ts
@@ -1,4 +1,0 @@
-import textEncoding = require("text-encoding");
-
-global.TextEncoder = textEncoding.TextEncoder;
-global.TextDecoder = textEncoding.TextDecoder;