+++ CHANGELOG.md
@@ -1,9 +1,23 @@
 # Next version
 
+## Breaking Changes
+
+- Move `accountIdentifierToBytes`, `accountIdentifierFromBytes` and `principalToAccountIdentifier` from `@dfinity/ledger-icp` to `@dfinity/nns`.
+- Rename `getLastestRewardEvent` to `getLatestRewardEvent` in `@dfinity/nns`.
+
 ## Features
 
 - Updated `@dfinity/nns` to add support to a new type of action `RegisterExtension`.
 - Updated `@dfinity/nns` to add support for setting following for multiple topics (`setFollowing`).
+- Updated `@dfinity/nns` to add support for get network metrics (`getMetrics`).
+
+## Build
+
+- Add explicit dependency requirement on `buffer` to `@dfinity/nns`.
+
+## Chore
+
+- Migrate from `Jest` to `Vitest`.
 
 # v70